[architect] refactor: split tools.go into domain-specific files
Split the 65KB tools.go (31 functions, 5 domains) into focused files
following the pattern already established in pkg/deploy/mcp/:

- tools_cluster.go:   cluster discovery and health (3 functions)
- tools_workloads.go: pod/deployment/service/node/event inspection (8 functions)
- tools_rbac.go:      role/permission/audit tools (11 functions)
- tools_policy.go:    Gatekeeper/ownership policy tools (6 functions)
- tools_drift.go:     GitOps drift detection (1 function)
- tools_helpers.go:   shared client factory functions (2 functions)

No behavioral changes — pure file reorganization within the same package.

Fixes #174
